More unsigned int to size_t conversion.

+++ auxiliary/gmp_matrix_add.c
@@ -18,29 +18,29 @@
 #include <sortsmill/math/gmp_matrix.h>
 
 VISIBLE void
-_GMP_TYPE (_matrix_add) (unsigned int m, unsigned int n,
+_GMP_TYPE (_matrix_add) (size_t m, size_t n,
                          _GMP_TYPE (_t) A[m][n], _GMP_TYPE (_t) B[m][n])
 {
-  for (unsigned int i = 0; i < m; i++)
-    for (unsigned int j = 0; j < n; j++)
+  for (size_t i = 0; i < m; i++)
+    for (size_t j = 0; j < n; j++)
       _GMP_TYPE (_add) (A[i][j], A[i][j], B[i][j]);
 }
 
 VISIBLE void
-_GMP_TYPE (_matrix_sub) (unsigned int m, unsigned int n,
+_GMP_TYPE (_matrix_sub) (size_t m, size_t n,
                          _GMP_TYPE (_t) A[m][n], _GMP_TYPE (_t) B[m][n])
 {
-  for (unsigned int i = 0; i < m; i++)
-    for (unsigned int j = 0; j < n; j++)
+  for (size_t i = 0; i < m; i++)
+    for (size_t j = 0; j < n; j++)
       _GMP_TYPE (_sub) (A[i][j], A[i][j], B[i][j]);
 }
 
 VISIBLE void
-_GMP_TYPE (_matrix_add_constant) (unsigned int m, unsigned int n,
+_GMP_TYPE (_matrix_add_constant) (size_t m, size_t n,
                                   _GMP_TYPE (_t) A[m][n],
                                   const _GMP_TYPE (_t) x)
 {
-  for (unsigned int i = 0; i < m; i++)
-    for (unsigned int j = 0; j < n; j++)
+  for (size_t i = 0; i < m; i++)
+    for (size_t j = 0; j < n; j++)
       _GMP_TYPE (_add) (A[i][j], A[i][j], x);
 }
